setting up behave and factory boy
