--- foodi/tests/features/login_steps.py
@@ -1,32 +1,0 @@
-from aloe import before, step, world
-from aloe.tools import guess_types
-from aloe_django.steps.models import get_model
-from nose.tools import assert_true
-from django.contrib.auth.models import User
-from rest_framework.test import APIClient
-
-
-@before.each_feature
-def before_each_feature(feature):
-    world.client = APIClient()
-
-
-@step('I empty the "([^"]+)" table')
-def step_empty_table(self, model_name):
-    get_model(model_name).objects.all().delete()
-
-
-@step('I create the following users:')
-def step_create_users(self):
-    for user in guess_types(self.hashes):
-        User.objects.create_user(**user)
-
-
-@step('I log in with username "([^"]+)" and password "([^"]+)"')
-def step_log_in(self, username, password):
-    world.is_logged_in = world.client.login(username=username, password=password)
-
-
-@step('I am logged in')
-def step_confirm_log_in(self):
-    assert_true(world.is_logged_in)
